my_crewai_project/pyproject.toml
[project]
name = "my-crewai-project"
version = "0.1.0"
description = "CrewAI agent system for content creation and analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.150.0",
    "python-dotenv>=1.0.0",
    "groq>=0.30.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
